Rework partition alignment
Moving the beginning of a partition creates a gap, and under some
circumstances fdisk tries to create the next partition inside this gap,
which then fails. Address this by moving the alignment after all
partitions have been created. This needs support in aii-ks, so use the
opportunity to let aii-ks emit the definition of some helper functions
too.