Introduce the should_create_ks method
There are two sections in the Kickstart file we need to add information
about block devices: the command section for Anaconda, and the %pre
script. Previously the should_print_ks method controlled both. To handle
ignored disks properly however, we must be able to supress printing
information in the command section, while still emitting code for the
%pre script.

For this reason, the should_create_ks method is introduced: it controls
emitting code in the %pre script, while should_print_ks now only
controls the Anaconda command section.