Fix caching of LVM and MD devices
The code previously only used the name of the device as the cache key.
However if aii-shellfe is told to configure two hosts that use the same
LVM volume, but the names of the physical volumes are different (e.g.
/dev/sda1 vs. /dev/cciss/c0d0p1), then one of the hosts will use
incorrect device names.

The fix is to generalize the caching changes made to Disk.pm earlier.