treat -1 same as 100% when defined as the end of a partition
There is a significant amount of legacy configuration that assumes that -1 means use all available free space when given as the end of the partition. The installation tools (parted) support this. It may be simpler to just get rid of the legacy code, but I am not sure how easy is this in sites that may have old 5.x systems still in use.