Implement lvmcache on lv volume functions and tests
This implements the creation of lvmcache lv volumes.
The original volume and cache volumes are created if not yet existing.
The volumes are then converted to a cached LV. When a LV of a cached LV is removed,
the cache is also automaticly removed by lvm.

This implements the code for components even as the ks commands