Add support for more Kickstart options on RH7
On RH7, "logvol" and "raid" also supports --label in addition to "part",
and --mkfsoptions is also supported everywhere.

To make use of this functionality, aii-ks needs to start passing the
Anaconda version when creating blockdevice objects. The alternative
approach, making ncm-lib-blockdevices parse the configuration of aii-ks
was more hackish and more problematic from the design angle.